generic-adapter/requirements-s3.txt
python-dotenv
telicent-lib==2.0.1
boto3